sandbox/examples/langchain-agent/pyproject.toml
[project]
name = "langchain-agent"
version = "0.1.0"
description = "LangChain agent with agent-sandbox integration using Volcengine API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agent-sandbox",
    "langchain>=1.0.0",
    "langchain-core>=1.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.2",
]

[tool.uv.sources]
agent-sandbox = { path = "../../sdk/python", editable = true }
